*) Fixed a type error in PerspectiveCamera's aspect ratio calculation *) Fixed a silly mistake in Camera w.r.t. returning polygon mode.

+++ ManagedModeller/PerspectiveCamera.cs
@@ -8,7 +8,7 @@ namespace ManagedModeller {
         public float GetFovY() { return fovY; }
         public void SetFovY(float fovY) { this.fovY = fovY; }
 
-        public float GetAspectRatio() { return width / height; }
+        public float GetAspectRatio() { return ((float) width) / height; }
 
         public PerspectiveCamera() {
             SetLocation(new Vector3(100, 100, 100));
